+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/classic/privateLinks/index.ts
@@ -17,7 +17,10 @@ export interface PrivateLinksOperations {
   ) => PagedAsyncIterableIterator<PrivateLinkResource>;
 }
 
-export function getPrivateLinks(context: HealthDataAIServicesContext, subscriptionId: string) {
+export function getPrivateLinks(
+  context: HealthDataAIServicesContext,
+  subscriptionId: string,
+) {
   return {
     listByDeidService: (
       resourceGroupName: string,