+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/healthDataAIServicesClient.ts
@@ -1,13 +1,22 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { getOperationsOperations, OperationsOperations } from "./classic/operations/index.js";
-import { getDeidServicesOperations, DeidServicesOperations } from "./classic/deidServices/index.js";
+import {
+  getOperationsOperations,
+  OperationsOperations,
+} from "./classic/operations/index.js";
+import {
+  getDeidServicesOperations,
+  DeidServicesOperations,
+} from "./classic/deidServices/index.js";
 import {
   getPrivateEndpointConnectionsOperations,
   PrivateEndpointConnectionsOperations,
 } from "./classic/privateEndpointConnections/index.js";
-import { getPrivateLinksOperations, PrivateLinksOperations } from "./classic/privateLinks/index.js";
+import {
+  getPrivateLinksOperations,
+  PrivateLinksOperations,
+} from "./classic/privateLinks/index.js";
 import {
   createHealthDataAIServices,
   HealthDataAIServicesContext,