+++ sdk/healthdataaiservices/arm-healthdataaiservices/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Release History
     
+## 2.0.0 (2024-12-05)
+    
+### Features Added
+
+  - Added Interface ErrorAdditionalInfo
+  - Added Interface ErrorDetail
+  - Added Interface ErrorResponse
+  - Enum KnownManagedServiceIdentityType has a new value "SystemAssigned,UserAssigned"
+  - Enum KnownOrigin has a new value "user,system"
+  - Enum KnownOrigin has a new value system
+  - Enum KnownOrigin has a new value user
+  - Enum KnownVersions has a new value v2024_09_20
+
+### Breaking Changes
+
+  - Enum KnownManagedServiceIdentityType no longer has value SystemAssignedUserAssigned
+  - Enum KnownOrigin no longer has value System
+  - Enum KnownOrigin no longer has value User
+  - Enum KnownOrigin no longer has value UserSystem
+  - Enum KnownVersions no longer has value V2024_09_20
+    
+    
 ## 1.0.0 (2024-11-25)
 
 ### Features Added